Change build type to Release and remove VRF test
The build configuration in 'ubuntu.yml' was changed from Debug to Release for better optimization and speed. An unnecessary VRF test, 'CHECK_MX_1', was also removed from 'vrf.cpp' to streamline the testing process. The test command in the 'ubuntu.yml' file was updated to include extra verbosity for better logs.