Update build command in Github Actions workflow
The build command in the Ubuntu Github Actions workflow has been updated. Instead of hardcoding the number of parallel jobs to 4, the build command now dynamically gets the number of logical CPUs available on the Github Actions runner using the 'sysctl' command. This allows for faster build times on runners with more than 4 CPUs.